# Conflicts:
#	src/main/java/com/example/back_end/controller/ProfitController.java
#	src/main/java/com/example/back_end/repository/IProfitRepository.java
#	src/main/java/com/example/back_end/service/IProfitService.java
#	src/main/java/com/example/back_end/service/ProfitServiceImpl.java
#	src/test/java/com/example/back_end/controller/ContractRestController_top10NewContract.java
#	src/test/java/com/example/back_end/controller/ProfitController_getAllContract.java
#	src/test/java/com/example/back_end/controller/redeemContractRestController/RedeemContractController_findOpenContractById.java
#	src/test/java/com/example/back_end/controller/redeemContractRestController/RedeemContractRestController_Redeem.java

+++ src/main/java/com/example/back_end/controller/ContractRestController.java
@@ -1,20 +1,35 @@
 package com.example.back_end.controller;
 
+import com.example.back_end.dto.CreateContractDto;
 import com.example.back_end.model.Contracts;
 import com.example.back_end.projections.ContractSearchDTO;
 import com.example.back_end.projections.IContractProjection;
 import com.example.back_end.service.IContractService;
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import java.util.List;
 import java.util.Optional;
 
+/**
+ * Created by: DinhHD
+ * Date created: 13/07/2023
+ * Function: do about pawn interface, customer selection interface
+ * <p>
+ * // * @param Contracts
+ *
+ * @return createContracts()
+ */
 
-@RestController
 @RequestMapping("/api/employee/contract")
+@RestController
+
 @CrossOrigin("*")
 public class ContractRestController {
     @Autowired
@@ -30,19 +45,16 @@ public class ContractRestController {
      * @return ResponseEntity<Page < IContractProjection>>
      */
 
-    @GetMapping("")
-    public ResponseEntity<Page<IContractProjection>> getALlTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
-                                                                              @RequestParam(name = "limit", defaultValue = "5") Integer limit) {
-        Page<IContractProjection> contractProjectionPage = this.iContractService.findAllTransactionHistory(page, limit);
-        int totalPage = contractProjectionPage.getTotalPages();
-        if (page >= totalPage) {
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
-        }
-        if (contractProjectionPage.isEmpty()) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-        }
-        return new ResponseEntity<>(contractProjectionPage, HttpStatus.OK);
-    }
+//    @GetMapping("")
+//    public ResponseEntity<Page<IContractProjection>> getAllTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
+//                                                                              @RequestParam(name = "limit", defaultValue = "5") Integer limit) {
+//        Page<IContractProjection> contractProjectionPage = this.iContractService.findAllTransactionHistory(page, limit);
+//        int totalPage = contractProjectionPage.getTotalPages();
+//        if (page >= totalPage) {
+//            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+//        }
+//        return new ResponseEntity<>(contractProjectionPage, HttpStatus.OK);
+//    }
 
     /**
      * Created by: ThienNT
@@ -55,7 +67,11 @@ public class ContractRestController {
      */
 
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Boolean> deleteTransactionHistoryById(@PathVariable("id") Integer id) {
+    public ResponseEntity<Boolean> deleteTransactionHistoryById(@PathVariable("id") Long id) {
+        Optional<Contracts> contractDTO = iContractService.findTransactionHistoryById(id);
+        if (!contractDTO.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         Boolean check = iContractService.deleteTransactionHistoryById(id);
         return new ResponseEntity<>(check, HttpStatus.OK);
     }
@@ -71,7 +87,7 @@ public class ContractRestController {
      */
 
     @GetMapping("/detail/{id}")
-    public ResponseEntity<Contracts> getTransactionHistoryDetail(@PathVariable("id") Integer id) {
+    public ResponseEntity<Contracts> showTransactionHistoryDetail(@PathVariable("id") Long id) {
         Optional<Contracts> contractDTO = iContractService.findTransactionHistoryById(id);
         return contractDTO.map(iContractProjection -> new ResponseEntity<>(iContractProjection, HttpStatus.OK)).orElseGet(() -> new ResponseEntity<>(HttpStatus.BAD_REQUEST));
     }
@@ -87,11 +103,32 @@ public class ContractRestController {
      * @requestbody contractSearchDTO
      */
 
-    @GetMapping("/search")
+    @GetMapping("")
     public ResponseEntity<Page<IContractProjection>> searchTransactionHistory(@RequestParam(name = "page", defaultValue = "0") Integer page,
                                                                               @RequestParam(name = "limit", defaultValue = "5") Integer limit,
                                                                               @RequestBody ContractSearchDTO contractSearchDTO) {
         Page<IContractProjection> contractProjectionsPage = iContractService.searchTransactionHistory(page, limit, contractSearchDTO);
+        int totalPage = contractProjectionsPage.getTotalPages();
+        if (page >= totalPage) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity<>(contractProjectionsPage, HttpStatus.OK);
+    }
+
+    @GetMapping("/list")
+    public ResponseEntity<List<Contracts>> getAllContract() {
+        List<Contracts> contractsList = iContractService.findAll();
+        return new ResponseEntity<>(contractsList, HttpStatus.OK);
+    }
+
+    @PostMapping("/createContract")
+    public ResponseEntity<?> createContracts(@RequestBody @Valid CreateContractDto contractDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        Contracts contracts = new Contracts();
+        BeanUtils.copyProperties(contractDto, contracts);
+        iContractService.createContract(contracts);
+        return new ResponseEntity<>(HttpStatus.CREATED);
     }
 }
